api/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "calendar-agent-api"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.111.0",
    "uvicorn[standard]==0.29.0",
    "sqlalchemy[asyncio]==2.0.30",
    "asyncpg==0.29.0",
    "alembic==1.13.1",
    "pydantic==2.7.1",
    "pydantic-settings==2.2.1",
    "httpx==0.27.0",
    "redis[hiredis]==5.0.4",
    "cryptography==42.0.7",
    "python-dateutil==2.9.0",
    "structlog==24.1.0",
    "slowapi==0.1.9",
    "calendar-agent-shared",
]

[tool.setuptools.packages.find]
where = ["."]
